+++ src/middleware/get-query.ts
@@ -29,7 +29,7 @@ export async function getQuery(
 	next?: NextFunction
 ) {
 	// Run model hook
-	if (!await runHook(request, response, 'getQuery', request.query)) {
+	if (!runHook(request, response, 'onGetQuery', request.query)) {
 		return;
 	}